Ticket #103: Add Win64 Support to the FDO API and SDF Provider
git-svn-id: http://svn.osgeo.org/fdo/trunk@3192 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Providers/SDF/Src/Provider/SdfCreateSDFFile.cpp
@@ -109,7 +109,7 @@ void SdfCreateSDFFile::Execute()
         throw FdoConnectionException::Create(NlsMsgGetMain(FDO_NLSID(SDFPROVIDER_30_CONNECTION_OPEN)));
 
     //check if the specified filename exists
-    int len = wcstombs(NULL, (FdoString*)m_filename, 0);
+    size_t len = wcstombs(NULL, (FdoString*)m_filename, 0);
     char* mbsName = new char[len + 1];
     wcstombs(mbsName, (FdoString*)m_filename, len+1);
